requirements.txt
pyscreenshot
sounddevice
requests
pynput
cryptography
python-dotenv
opencv-python
